test: Revert in-source testing
