+++ src/lib/chrome/bookmark.test.ts
@@ -1,0 +1,6 @@
+import { describe, it } from 'vitest';
+import { createBookmarks } from './bookmark';
+
+describe(createBookmarks, () => {
+	it.todo('creates bookmark recursively');
+});
