+++ src/lib/chrome/extension.ts
@@ -34,29 +34,3 @@ export function getExtensionID(): string {
 
 	return extensionID;
 }
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { default: path } = await import('path');
-	const { describe, expect, it } = import.meta.vitest;
-
-	describe(getProjectRoot, () => {
-		it('should be', () => {
-			expect(getProjectRoot()).toEqual(path.join(__dirname, '../../../'));
-		});
-	});
-
-	describe(getExtensionPath, () => {
-		it('should be `dist` directory in project root', async () => {
-			expect(getExtensionPath()).toEqual(path.join(getProjectRoot(), 'dist'));
-		});
-	});
-
-	describe(getExtensionID, () => {
-		// NOTE: Validity should be checked at e2e tests as it requires Chrome browser behavior, the actual extension ID
-		it('should be 32 length string of lowercase alphabets', () => {
-			expect(getExtensionID()).toMatch(/[a-z]{32}/);
-		});
-	});
-}
-/* c8 ignore stop */
