+++ src/lib/messages.test.ts
@@ -1,0 +1,3 @@
+import { it } from 'vitest';
+
+it.todo('nothing to test yet');
