+++ src/lib/raindrop/client.ts
@@ -68,11 +68,3 @@ export function getAxiosClient(): AxiosInstance {
 
 	return instance;
 }
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { it } = import.meta.vitest;
-
-	it.todo('nothing to test yet');
-}
-/* c8 ignore stop */
