+++ src/lib/raindrop/index.ts
@@ -7,11 +7,3 @@
 import { getClient } from './client';
 
 export default getClient();
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { it } = import.meta.vitest;
-
-	it.todo('nothing to test yet');
-}
-/* c8 ignore stop */
