+++ src/manifest.ts
@@ -21,15 +21,3 @@ const manifest = {
 };
 
 export default manifest;
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { expect, it } = import.meta.vitest;
-
-	it('metadata should match project info', () => {
-		expect(manifest.name).toEqual('Raindrop Sync for Chrome');
-		expect(manifest.version).toEqual(packageManifest.version);
-		expect(manifest.description).not.toEqual('');
-	});
-}
-/* c8 ignore stop */
