+++ src/service-worker.ts
@@ -8,11 +8,3 @@ chrome.runtime.onInstalled.addListener((details) => {
 			break;
 	}
 });
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { it } = import.meta.vitest;
-
-	it.todo('nothing to test yet');
-}
-/* c8 ignore stop */
